reorganize lending pallet
