+++ frame/lending/src/impls/mod.rs
@@ -1,0 +1,1 @@
+mod lending;
