+++ frame/lending/src/lib.rs
@@ -42,6 +42,9 @@ pub mod validation;
 pub mod weights;
 pub use crate::weights::WeightInfo;
 
+pub mod crypto;
+mod helpers;
+mod impls;
 mod models;
 
 #[cfg(test)]
@@ -62,37 +65,24 @@ mod repay_borrow;
 
 #[frame_support::pallet]
 pub mod pallet {
-	use crate::{models::borrower_data::BorrowerData, weights::WeightInfo};
+	use crate::weights::WeightInfo;
 
-	use crate::validation::{
-		AssetIsSupportedByOracle, BalanceGreaterThenZero, CurrencyPairIsNotSame, MarketModelValid,
-		UpdateInputValid,
-	};
+	pub use crate::helpers::swap;
 	use codec::Codec;
-	use composable_support::{
-		math::safe::{SafeAdd, SafeDiv, SafeMul, SafeSub},
-		validation::{TryIntoValidated, Validated},
-	};
 	use composable_traits::{
 		currency::CurrencyFactory,
 		defi::{DeFiComposableConfig, *},
-		lending::{
-			math::{self, *},
-			BorrowAmountOf, CollateralLpAmountOf, CreateInput, Lending, MarketConfig,
-			RepayStrategy, TotalDebtWithInterest, UpdateInput,
-		},
+		lending::{CreateInput, Lending, MarketConfig, RepayStrategy, UpdateInput},
 		liquidation::Liquidation,
 		oracle::Oracle,
-		time::{DurationSeconds, Timestamp, SECONDS_PER_YEAR_NAIVE},
-		vault::{Deposit, FundsAvailability, StrategicVault, Vault, VaultConfig},
+		time::Timestamp,
+		vault::StrategicVault,
 	};
 	use frame_support::{
 		pallet_prelude::*,
-		storage::{with_transaction, TransactionOutcome},
 		traits::{
 			fungible::{Inspect as NativeInspect, Transfer as NativeTransfer},
-			fungibles::{Inspect, InspectHold, Mutate, MutateHold, Transfer},
-			tokens::DepositConsequence,
+			fungibles::{InspectHold, Mutate, MutateHold, Transfer},
 			UnixTime,
 		},
 		transactional,
@@ -104,14 +94,11 @@ pub mod pallet {
 		pallet_prelude::*,
 	};
 	use sp_core::crypto::KeyTypeId;
-	use sp_runtime::{
-		traits::{AccountIdConversion, One, Saturating, Zero},
-		ArithmeticError, DispatchError, FixedPointNumber, FixedU128, KeyTypeId as CryptoKeyTypeId,
-		Percent, Perquintill,
-	};
+	use sp_runtime::KeyTypeId as CryptoKeyTypeId;
 	use sp_std::{fmt::Debug, vec, vec::Vec};
+
 	/// Simple type alias around [`MarketConfig`] for this pallet.
-	type MarketConfigOf<T> = MarketConfig<
+	pub(crate) type MarketConfigOf<T> = MarketConfig<
 		<T as Config>::VaultId,
 		<T as DeFiComposableConfig>::MayBeAssetId,
 		<T as frame_system::Config>::AccountId,
@@ -144,47 +131,10 @@ pub mod pallet {
 		pub(crate) debt_asset: <T as DeFiComposableConfig>::MayBeAssetId,
 	}
 
-	/// Used to count the calls in [`Pallet::initialize_block`]. Each field corresponds to a
-	/// function call to count.
-	#[derive(Debug, Default, Clone, Copy)]
-	pub(crate) struct InitializeBlockCallCounters {
-		now: u32,
-		read_markets: u32,
-		accrue_interest: u32,
-		account_id: u32,
-		available_funds: u32,
-		handle_withdrawable: u32,
-		handle_depositable: u32,
-		handle_must_liquidate: u32,
-	}
-
 	pub const KEY_TYPE: KeyTypeId = KeyTypeId(*b"lend");
 	pub const CRYPTO_KEY_TYPE: CryptoKeyTypeId = CryptoKeyTypeId(*b"lend");
 
-	pub mod crypto {
-		use super::KEY_TYPE;
-		use frame_system::offchain;
-		use sp_core::sr25519::{self, Signature as Sr25519Signature};
-		use sp_runtime::{app_crypto::app_crypto, traits::Verify, MultiSignature, MultiSigner};
-
-		app_crypto!(sr25519, KEY_TYPE);
-
-		pub struct TestAuthId;
-
-		// implementation for runtime
-		impl offchain::AppCrypto<MultiSigner, MultiSignature> for TestAuthId {
-			type RuntimeAppPublic = Public;
-			type GenericSignature = sr25519::Signature;
-			type GenericPublic = sr25519::Public;
-		}
-
-		// implementation for mock runtime in test
-		impl offchain::AppCrypto<<Sr25519Signature as Verify>::Signer, Sr25519Signature> for TestAuthId {
-			type RuntimeAppPublic = Public;
-			type GenericSignature = sr25519::Signature;
-			type GenericPublic = sr25519::Public;
-		}
-	}
+	pub use crate::crypto;
 
 	#[pallet::config]
 	pub trait Config:
@@ -775,1178 +725,9 @@ pub mod pallet {
 		}
 	}
 
-	// public helper functions
-	impl<T: Config> Pallet<T> {
-		/// Returns the initial pool size for a market with `borrow_asset`. Calculated with
-		/// [`Config::OracleMarketCreationStake`].
-		pub fn calculate_initial_pool_size(
-			borrow_asset: <T::Oracle as composable_traits::oracle::Oracle>::AssetId,
-		) -> Result<<T as composable_traits::defi::DeFiComposableConfig>::Balance, DispatchError> {
-			T::Oracle::get_price_inverse(borrow_asset, T::OracleMarketCreationStake::get())
-		}
-
-		/// Creates a new [`BorrowerData`] for the given market and account. See [`BorrowerData`]
-		/// for more information.
-		pub fn create_borrower_data(
-			market_id: &<Self as Lending>::MarketId,
-			account: &<Self as DeFiEngine>::AccountId,
-		) -> Result<BorrowerData, DispatchError> {
-			let (_, market) = Self::get_market(market_id)?;
-
-			let collateral_balance_value = Self::get_price(
-				market.collateral_asset,
-				Self::collateral_of_account(market_id, account)?,
-			)?;
-
-			let account_total_debt_with_interest =
-				Self::total_debt_with_interest(market_id, account)?.unwrap_or_zero();
-			let borrow_balance_value = Self::get_price(
-				T::Vault::asset_id(&market.borrow_asset_vault)?,
-				account_total_debt_with_interest,
-			)?;
-
-			let borrower =
-				BorrowerData::new(
-					collateral_balance_value,
-					borrow_balance_value,
-					market
-						.collateral_factor
-						.try_into_validated()
-						.map_err(|_| Error::<T>::CollateralFactorMustBeMoreThanOne)?, /* TODO: Use a proper error mesage */
-					market.under_collateralized_warn_percent,
-				);
-
-			Ok(borrower)
-		}
-
-		/// Whether or not an account should be liquidated. See [`BorrowerData::should_liquidate()`]
-		/// for more information.
-		pub fn should_liquidate(
-			market_id: &<Self as Lending>::MarketId,
-			account: &<Self as DeFiEngine>::AccountId,
-		) -> Result<bool, DispatchError> {
-			let borrower = Self::create_borrower_data(market_id, account)?;
-			let should_liquidate = borrower.should_liquidate()?;
-			Ok(should_liquidate)
-		}
-
-		pub fn soon_under_collateralized(
-			market_id: &<Self as Lending>::MarketId,
-			account: &<Self as DeFiEngine>::AccountId,
-		) -> Result<bool, DispatchError> {
-			let borrower = Self::create_borrower_data(market_id, account)?;
-			let should_warn = borrower.should_warn()?;
-			Ok(should_warn)
-		}
-
-		/// Initiate liquidation of individual position for particular borrower within mentioned
-		/// market. Returns 'Ok(())' in the case of successful initiation, 'Err(DispatchError)' in
-		/// the opposite case.
-		/// - `liquidator` : Liquidator's account id.
-		/// - `market_pair` : Index and configuration of the market from which tokens were borrowed.
-		/// - `account` : Borrower's account id whose debt are going to be liquidated.
-		fn liquidate_position(
-			liquidator: &<Self as DeFiEngine>::AccountId,
-			market_pair: &(&<Self as Lending>::MarketId, MarketConfigOf<T>),
-			borrow_asset: <T as DeFiComposableConfig>::MayBeAssetId,
-			account: &<Self as DeFiEngine>::AccountId,
-		) -> Result<(), DispatchError> {
-			let (market_id, market) = market_pair;
-			ensure!(
-				Self::should_liquidate(market_id, account)?,
-				DispatchError::Other(
-					"Tried liquidate position which is not supposed to be liquidated"
-				)
-			);
-
-			let collateral_to_liquidate = Self::collateral_of_account(market_id, account)?;
-
-			let source_target_account = Self::account_id(market_id);
-
-			let unit_price =
-				T::Oracle::get_ratio(CurrencyPair::new(market.collateral_asset, borrow_asset))?;
-
-			let sell = Sell::new(
-				market.collateral_asset,
-				borrow_asset,
-				collateral_to_liquidate,
-				unit_price,
-			);
-			T::Liquidation::liquidate(&source_target_account, sell, market.liquidators.clone())?;
-			if let Some(deposit) = BorrowRent::<T>::get(market_id, account) {
-				let market_account = Self::account_id(market_id);
-				<T as Config>::NativeCurrency::transfer(
-					&market_account,
-					liquidator,
-					deposit,
-					false,
-				)?;
-			}
-			Ok(())
-		}
-
-		/// Liquidates debt for each borrower in the vector within mentioned market.
-		/// Returns a vector of borrowers' account ids whose debts were liquidated.
-		/// - `liquidator` : Liquidator's account id.
-		/// - `market_id` : Market index from which `borrowers` has taken borrow.
-		/// - `borrowers` : Vector of borrowers whose debts are going to be liquidated.
-		pub fn liquidate_internal(
-			liquidator: &<Self as DeFiEngine>::AccountId,
-			market_id: &<Self as Lending>::MarketId,
-			borrowers: BoundedVec<<Self as DeFiEngine>::AccountId, T::MaxLiquidationBatchSize>,
-		) -> Result<Vec<<Self as DeFiEngine>::AccountId>, DispatchError> {
-			// Vector of borrowers whose positions are involved in the liquidation process.
-			let mut subjected_borrowers: Vec<<Self as DeFiEngine>::AccountId> = Vec::new();
-			let market_pair = Self::get_market(market_id)?;
-			let borrow_asset = T::Vault::asset_id(&market_pair.1.borrow_asset_vault)?;
-			for account in borrowers.iter() {
-				// Wrap liquidate position request in a storage transaction.
-				// So, in the case of any error state's changes will not be committed
-				let storage_transaction_succeeded = with_transaction(|| {
-					let liquidation_response_result =
-						Self::liquidate_position(liquidator, &market_pair, borrow_asset, account);
-					if let Err(error) = liquidation_response_result {
-						log::warn!("Creation of liquidation request for position {:?} {:?} was failed: {:?}",
-							market_id,
-							account,
-							error );
-						return TransactionOutcome::Rollback(liquidation_response_result)
-					}
-					TransactionOutcome::Commit(Ok(()))
-				});
-
-				// If storage transaction succeeded,
-				// push borrower to the output vector,
-				// remove debt records from storages.
-				if storage_transaction_succeeded.is_ok() {
-					subjected_borrowers.push(account.clone());
-					BorrowTimestamp::<T>::remove(market_id, account);
-					DebtIndex::<T>::remove(market_id, account);
-				}
-			}
-			Ok(subjected_borrowers)
-		}
-	}
-
-	// crate-public helper functions
-	impl<T: Config> Pallet<T> {
-		pub(crate) fn initialize_block(
-			block_number: T::BlockNumber,
-		) -> InitializeBlockCallCounters {
-			let mut call_counters = InitializeBlockCallCounters::default();
-			let _ = with_transaction(|| {
-				let now = Self::now();
-				call_counters.now += 1;
-
-				let mut errors = Markets::<T>::iter()
-					.map(|(market_id, config)| {
-						call_counters.read_markets += 1;
-						Self::accrue_interest(&market_id, now)?;
-						call_counters.accrue_interest += 1;
-						let market_account = Self::account_id(&market_id);
-						call_counters.account_id += 1;
-						// NOTE(hussein-aitlahcen):
-						// It would probably be more perfomant to handle theses
-						// case while borrowing/repaying.
-						//
-						// I don't know whether we would face any issue by doing that.
-						//
-						// borrow:
-						//  - withdrawable = transfer(vault->market) + transfer(market->user)
-						//  - depositable = error(not enough borrow asset) // vault asking for
-						//    reserve to be fullfilled
-						//  - mustliquidate = error(market is closing)
-						// repay:
-						// 	- (withdrawable || depositable || mustliquidate) =
-						//    transfer(user->market) + transfer(market->vault)
-						//
-						// The intermediate transfer(vault->market) while borrowing would
-						// allow the vault to update the strategy balance (market = borrow vault
-						// strategy).
-						match Self::available_funds(&config, &market_account)? {
-							FundsAvailability::Withdrawable(balance) => {
-								Self::handle_withdrawable(&config, &market_account, balance)?;
-								call_counters.handle_withdrawable += 1;
-							},
-							FundsAvailability::Depositable(balance) => {
-								Self::handle_depositable(&config, &market_account, balance)?;
-								call_counters.handle_depositable += 1;
-							},
-							FundsAvailability::MustLiquidate => {
-								Self::handle_must_liquidate(&config, &market_account)?;
-								call_counters.handle_must_liquidate += 1;
-							},
-						}
-
-						call_counters.available_funds += 1;
-
-						Result::<(), DispatchError>::Ok(())
-					})
-					.filter_map(|r| match r {
-						Ok(_) => None,
-						Err(err) => Some(err),
-					})
-					.peekable();
-
-				if errors.peek().is_none() {
-					LastBlockTimestamp::<T>::put(now);
-					TransactionOutcome::Commit(Ok(1000))
-				} else {
-					errors.for_each(|e| {
-						log::error!(
-							"This should never happen, could not initialize block!!! {:#?} {:#?}",
-							block_number,
-							e
-						)
-					});
-					TransactionOutcome::Rollback(Err(DispatchError::Other(
-						"failed to initialize block",
-					)))
-				}
-			});
-			call_counters
-		}
-
-		pub(crate) fn now() -> u64 {
-			T::UnixTime::now().as_secs()
-		}
-
-		pub(crate) fn available_funds(
-			config: &MarketConfigOf<T>,
-			market_account: &T::AccountId,
-		) -> Result<FundsAvailability<T::Balance>, DispatchError> {
-			<T::Vault as StrategicVault>::available_funds(
-				&config.borrow_asset_vault,
-				market_account,
-			)
-		}
-
-		pub(crate) fn handle_withdrawable(
-			config: &MarketConfigOf<T>,
-			market_account: &T::AccountId,
-			balance: T::Balance,
-		) -> Result<(), DispatchError> {
-			<T::Vault as StrategicVault>::withdraw(
-				&config.borrow_asset_vault,
-				market_account,
-				balance,
-			)
-		}
-
-		pub(crate) fn handle_depositable(
-			config: &MarketConfigOf<T>,
-			market_account: &T::AccountId,
-			balance: T::Balance,
-		) -> Result<(), DispatchError> {
-			let asset_id = <T::Vault as Vault>::asset_id(&config.borrow_asset_vault)?;
-			let balance =
-				<T as Config>::MultiCurrency::reducible_balance(asset_id, market_account, false)
-					.min(balance);
-			<T::Vault as StrategicVault>::deposit(
-				&config.borrow_asset_vault,
-				market_account,
-				balance,
-			)
-		}
-
-		pub(crate) fn handle_must_liquidate(
-			config: &MarketConfigOf<T>,
-			market_account: &T::AccountId,
-		) -> Result<(), DispatchError> {
-			let asset_id = <T::Vault as Vault>::asset_id(&config.borrow_asset_vault)?;
-			let balance =
-				<T as Config>::MultiCurrency::reducible_balance(asset_id, market_account, false);
-			<T::Vault as StrategicVault>::deposit(
-				&config.borrow_asset_vault,
-				market_account,
-				balance,
-			)
-		}
-
-		/// Returns the borrow and debt assets for the given market, if it exists.
-		pub(crate) fn get_assets_for_market(
-			market_id: &MarketIndex,
-		) -> Result<MarketAssets<T>, DispatchError> {
-			let (_, market) = Self::get_market(market_id)?;
-			let borrow_asset = T::Vault::asset_id(&market.borrow_asset_vault)?;
-			let debt_asset =
-				DebtTokenForMarket::<T>::get(market_id).ok_or(Error::<T>::MarketDoesNotExist)?;
-
-			Ok(MarketAssets { borrow_asset, debt_asset })
-		}
-	}
-
-	// private helper functions
-	impl<T: Config> Pallet<T> {
-		fn do_create_market(
-			manager: T::AccountId,
-			input: Validated<
-				CreateInputOf<T>,
-				(MarketModelValid, CurrencyPairIsNotSame, AssetIsSupportedByOracle<T::Oracle>),
-			>,
-			keep_alive: bool,
-		) -> Result<(<Self as Lending>::MarketId, T::VaultId), DispatchError> {
-			let config_input = input.value();
-			LendingCount::<T>::try_mutate(|MarketIndex(previous_market_index)| {
-				let market_id = {
-					// TODO: early mutation of `previous_market_index` value before check.
-					*previous_market_index += 1;
-					ensure!(
-						*previous_market_index <= T::MaxMarketCount::get(),
-						Error::<T>::ExceedLendingCount
-					);
-					MarketIndex(*previous_market_index)
-				};
-
-				let borrow_asset_vault = T::Vault::create(
-					Deposit::Existential,
-					VaultConfig {
-						asset_id: config_input.borrow_asset(),
-						reserved: config_input.reserved_factor(),
-						manager: manager.clone(),
-						strategies: [(
-							Self::account_id(&market_id),
-							// Borrowable = 100% - reserved
-							// REVIEW: Review use of `saturating_sub` here - I'm pretty sure this
-							// can never error, but if `Perquintill` can be `>`
-							// `Perquintill::one()` then we might want to re-evaluate the logic
-							// here.
-							Perquintill::one().saturating_sub(config_input.reserved_factor()),
-						)]
-						.into_iter()
-						.collect(),
-					},
-				)?;
-
-				let initial_pool_size =
-					Self::calculate_initial_pool_size(config_input.borrow_asset())?;
-
-				ensure!(
-					initial_pool_size > T::Balance::zero(),
-					Error::<T>::PriceOfInitialBorrowVaultShouldBeGreaterThanZero
-				);
-
-				// transfer `initial_pool_size` worth of borrow asset from the manager to the market
-				T::MultiCurrency::transfer(
-					config_input.borrow_asset(),
-					&manager,
-					&Self::account_id(&market_id),
-					initial_pool_size,
-					keep_alive,
-				)?;
-
-				let market_config = MarketConfig {
-					manager,
-					max_price_age: config_input.updatable.max_price_age,
-					borrow_asset_vault: borrow_asset_vault.clone(),
-					collateral_asset: config_input.collateral_asset(),
-					collateral_factor: config_input.updatable.collateral_factor,
-					interest_rate_model: config_input.interest_rate_model,
-					under_collateralized_warn_percent: config_input
-						.updatable
-						.under_collateralized_warn_percent,
-					liquidators: config_input.updatable.liquidators,
-				};
-				// TODO: pass ED from API,
-				let debt_token_id = T::CurrencyFactory::reserve_lp_token_id(T::Balance::default())?;
-
-				DebtTokenForMarket::<T>::insert(market_id, debt_token_id);
-				Markets::<T>::insert(market_id, market_config);
-				BorrowIndex::<T>::insert(market_id, FixedU128::one());
-
-				Ok((market_id, borrow_asset_vault))
-			})
-		}
-		fn do_deposit_collateral(
-			market_id: &<Self as Lending>::MarketId,
-			account: &T::AccountId,
-			amount: Validated<CollateralLpAmountOf<Self>, BalanceGreaterThenZero>,
-			keep_alive: bool,
-		) -> Result<(), DispatchError> {
-			let amount = amount.value();
-			let (_, market) = Self::get_market(market_id)?;
-			let market_account = Self::account_id(market_id);
-
-			AccountCollateral::<T>::try_mutate(market_id, account, |collateral_balance| {
-				let new_collateral_balance =
-					collateral_balance.unwrap_or_default().safe_add(&amount)?;
-				collateral_balance.replace(new_collateral_balance);
-				Result::<(), DispatchError>::Ok(())
-			})?;
-
-			<T as Config>::MultiCurrency::transfer(
-				market.collateral_asset,
-				account,
-				&market_account,
-				amount,
-				keep_alive,
-			)?;
-			Ok(())
-		}
-
-		fn do_update_market(
-			manager: T::AccountId,
-			market_id: MarketIndex,
-			input: Validated<
-				UpdateInput<T::LiquidationStrategyId, <T as frame_system::Config>::BlockNumber>,
-				UpdateInputValid,
-			>,
-		) -> DispatchResultWithPostInfo {
-			let input = input.value();
-			Markets::<T>::mutate(&market_id, |market| {
-				if let Some(market) = market {
-					ensure!(manager == market.manager, Error::<T>::Unauthorized);
-
-					ensure!(
-						market.collateral_factor >= input.collateral_factor,
-						Error::<T>::CannotIncreaseCollateralFactorOfOpenMarket
-					);
-					market.collateral_factor = input.collateral_factor;
-					market.under_collateralized_warn_percent =
-						input.under_collateralized_warn_percent;
-					market.liquidators = input.liquidators.clone();
-					Ok(())
-				} else {
-					Err(Error::<T>::MarketDoesNotExist)
-				}
-			})?;
-			Self::deposit_event(Event::<T>::MarketUpdated { market_id, input });
-			Ok(().into())
-		}
-
-		/// Returns pair of market's id and market (as 'MarketConfing') via market's id
-		/// - `market_id` : Market index as a key in 'Markets' storage
-		fn get_market(
-			market_id: &MarketIndex,
-		) -> Result<(&MarketIndex, MarketConfigOf<T>), DispatchError> {
-			Markets::<T>::get(market_id)
-				.map(|market| (market_id, market))
-				.ok_or_else(|| Error::<T>::MarketDoesNotExist.into())
-		}
-
-		/// Get TWAP from oracle. If history of prices is empty then return latest price.
-		fn get_price(
-			asset_id: <T as DeFiComposableConfig>::MayBeAssetId,
-			amount: T::Balance,
-		) -> Result<T::Balance, DispatchError> {
-			<T::Oracle as Oracle>::get_twap_for_amount(asset_id, amount)
-		}
-
-		/// Some of these checks remain to provide better errors. See [this clickup task](task) for
-		/// more information.
-		///
-		/// [task]: <https://sharing.clickup.com/20465559/t/h/27yd3wt/7IB0QYYHXP0TZZT>
-		fn can_borrow(
-			market_id: &MarketIndex,
-			debt_owner: &T::AccountId,
-			amount_to_borrow: BorrowAmountOf<Self>,
-			market: MarketConfigOf<T>,
-			market_account: &T::AccountId,
-		) -> Result<(), DispatchError> {
-			// this check prevents free flash loans
-			if let Some(latest_borrow_timestamp) = BorrowTimestamp::<T>::get(market_id, debt_owner)
-			{
-				if latest_borrow_timestamp >= LastBlockTimestamp::<T>::get() {
-					return Err(Error::<T>::InvalidTimestampOnBorrowRequest.into())
-				}
-			}
-
-			let borrow_asset = T::Vault::asset_id(&market.borrow_asset_vault)?;
-			let borrow_limit = Self::get_borrow_limit(market_id, debt_owner)?;
-			let borrow_amount_value = Self::get_price(borrow_asset, amount_to_borrow)?;
-			ensure!(borrow_limit >= borrow_amount_value, Error::<T>::NotEnoughCollateralToBorrow);
-
-			ensure!(
-				<T as Config>::MultiCurrency::can_withdraw(
-					borrow_asset,
-					market_account,
-					amount_to_borrow
-				)
-				.into_result()
-				.is_ok(),
-				Error::<T>::NotEnoughBorrowAsset,
-			);
-
-			if !BorrowRent::<T>::contains_key(market_id, debt_owner) {
-				let deposit = T::WeightToFee::calc(&T::WeightInfo::liquidate(1));
-				// See note 1
-				ensure!(
-					<T as Config>::NativeCurrency::can_withdraw(debt_owner, deposit)
-						.into_result()
-						.is_ok(),
-					Error::<T>::NotEnoughRent,
-				);
-			}
-
-			ensure!(
-				!matches!(
-					T::Vault::available_funds(&market.borrow_asset_vault, market_account)?,
-					FundsAvailability::MustLiquidate
-				),
-				Error::<T>::MarketIsClosing
-			);
-
-			Ok(())
-		}
-
-		/// Check is price actual yet
-		fn ensure_price_is_recent(market: &MarketConfigOf<T>) -> Result<(), DispatchError> {
-			use sp_runtime::traits::CheckedSub as _;
-
-			let borrow_asset = T::Vault::asset_id(&market.borrow_asset_vault)?;
-
-			let current_block = frame_system::Pallet::<T>::block_number();
-			let blocks_count = market.max_price_age;
-			let edge_block = current_block.checked_sub(&blocks_count).unwrap_or_default();
-
-			// check borrow asset
-			let price_block =
-				<T::Oracle as Oracle>::get_price(borrow_asset, BorrowAmountOf::<Self>::default())?
-					.block;
-			ensure!(price_block >= edge_block, Error::<T>::PriceTooOld);
-
-			// check collateral asset
-			let collateral_asset = market.collateral_asset;
-			let price_block = <T::Oracle as Oracle>::get_price(
-				collateral_asset,
-				BorrowAmountOf::<Self>::default(),
-			)?
-			.block;
-			ensure!(price_block >= edge_block, Error::<T>::PriceTooOld);
-
-			Ok(())
-		}
-	}
-
 	impl<T: Config> DeFiEngine for Pallet<T> {
 		type MayBeAssetId = <T as DeFiComposableConfig>::MayBeAssetId;
 		type Balance = <T as DeFiComposableConfig>::Balance;
 		type AccountId = <T as frame_system::Config>::AccountId;
-	}
-
-	impl<T: Config> Lending for Pallet<T> {
-		type VaultId = <T::Vault as Vault>::VaultId;
-		type MarketId = MarketIndex;
-		type BlockNumber = T::BlockNumber;
-		type LiquidationStrategyId = <T as Config>::LiquidationStrategyId;
-		type Oracle = T::Oracle;
-
-		fn create_market(
-			manager: Self::AccountId,
-			input: CreateInputOf<T>,
-			keep_alive: bool,
-		) -> Result<(Self::MarketId, Self::VaultId), DispatchError> {
-			Self::do_create_market(manager, input.try_into_validated()?, keep_alive)
-		}
-
-		fn update_market(
-			manager: Self::AccountId,
-			market_id: Self::MarketId,
-			input: UpdateInput<Self::LiquidationStrategyId, Self::BlockNumber>,
-		) -> DispatchResultWithPostInfo {
-			Self::do_update_market(manager, market_id, input.try_into_validated()?)
-		}
-
-		fn account_id(market_id: &Self::MarketId) -> Self::AccountId {
-			T::PalletId::get().into_sub_account(market_id)
-		}
-
-		fn deposit_collateral(
-			market_id: &Self::MarketId,
-			account: &Self::AccountId,
-			amount: CollateralLpAmountOf<Self>,
-			keep_alive: bool,
-		) -> Result<(), DispatchError> {
-			Self::do_deposit_collateral(
-				market_id,
-				account,
-				amount.try_into_validated()?,
-				keep_alive,
-			)
-		}
-
-		fn withdraw_collateral(
-			market_id: &Self::MarketId,
-			account: &Self::AccountId,
-			amount: CollateralLpAmountOf<Self>,
-		) -> Result<(), DispatchError> {
-			let (_, market) = Self::get_market(market_id)?;
-
-			let collateral_balance = AccountCollateral::<T>::try_get(market_id, account)
-				// REVIEW: Perhaps don't default to zero
-				// REVIEW: What is expected behaviour if there is no collateral?
-				.unwrap_or_else(|_| CollateralLpAmountOf::<Self>::zero());
-
-			ensure!(amount <= collateral_balance, Error::<T>::NotEnoughCollateralToWithdraw);
-
-			let borrow_asset = T::Vault::asset_id(&market.borrow_asset_vault)?;
-			let borrower_balance_with_interest =
-				Self::total_debt_with_interest(market_id, account)?.unwrap_or_zero();
-
-			let borrow_balance_value =
-				Self::get_price(borrow_asset, borrower_balance_with_interest)?;
-
-			let collateral_balance_after_withdrawal_value =
-				Self::get_price(market.collateral_asset, collateral_balance.safe_sub(&amount)?)?;
-
-			let borrower_after_withdrawal = BorrowerData::new(
-				collateral_balance_after_withdrawal_value,
-				borrow_balance_value,
-				market
-					.collateral_factor
-					.try_into_validated()
-					.map_err(|_| Error::<T>::Overflow)?, // TODO: Use a proper error mesage?
-				market.under_collateralized_warn_percent,
-			);
-
-			ensure!(
-				!borrower_after_withdrawal.should_liquidate()?,
-				Error::<T>::WouldGoUnderCollateralized
-			);
-
-			let market_account = Self::account_id(market_id);
-
-			ensure!(
-				<T as Config>::MultiCurrency::can_deposit(
-					market.collateral_asset,
-					account,
-					amount,
-					false
-				) == DepositConsequence::Success,
-				Error::<T>::TransferFailed
-			);
-			ensure!(
-				<T as Config>::MultiCurrency::can_withdraw(
-					market.collateral_asset,
-					&market_account,
-					amount
-				)
-				.into_result()
-				.is_ok(),
-				Error::<T>::TransferFailed
-			);
-
-			AccountCollateral::<T>::try_mutate(market_id, account, |collateral_balance| {
-				let new_collateral_balance =
-					// REVIEW: Should we default if there's no collateral? Or should an error (something like "NoCollateralToWithdraw") be returned instead?
-					collateral_balance.unwrap_or_default().safe_sub(&amount)?;
-
-				collateral_balance.replace(new_collateral_balance);
-
-				Result::<(), DispatchError>::Ok(())
-			})?;
-			<T as Config>::MultiCurrency::transfer(
-				market.collateral_asset,
-				&market_account,
-				account,
-				amount,
-				true,
-			)
-			.expect("impossible; qed;");
-			Ok(())
-		}
-
-		fn get_markets_for_borrow(borrow: Self::VaultId) -> Vec<Self::MarketId> {
-			Markets::<T>::iter()
-				.filter_map(|(index, market)| market.borrow_asset_vault.eq(&borrow).then(|| index))
-				.collect()
-		}
-
-		fn borrow(
-			market_id: &Self::MarketId,
-			borrowing_account: &Self::AccountId,
-			amount_to_borrow: BorrowAmountOf<Self>,
-		) -> Result<(), DispatchError> {
-			let (_, market) = Self::get_market(market_id)?;
-
-			Self::ensure_price_is_recent(&market)?;
-
-			let MarketAssets { borrow_asset, debt_asset: debt_asset_id } =
-				Self::get_assets_for_market(market_id)?;
-
-			let market_account = Self::account_id(market_id);
-
-			Self::can_borrow(
-				market_id,
-				borrowing_account,
-				amount_to_borrow,
-				market,
-				&market_account,
-			)?;
-
-			let new_account_interest_index = {
-				let market_index =
-					BorrowIndex::<T>::get(market_id).ok_or(Error::<T>::MarketDoesNotExist)?;
-
-				// previous account interest index
-				let account_interest_index = DebtIndex::<T>::get(market_id, borrowing_account)
-					.unwrap_or_else(ZeroToOneFixedU128::zero);
-
-				// amount of debt currently
-				let existing_principal_amount =
-					<T as Config>::MultiCurrency::balance(debt_asset_id, borrowing_account);
-
-				// principal_after_new_borrow
-				let principal_after_new_borrow =
-					existing_principal_amount.safe_add(&amount_to_borrow)?;
-
-				// amount of principal the account already has
-				let existing_borrow_share =
-					Percent::from_rational(existing_principal_amount, principal_after_new_borrow);
-				// amount of principal the account is adding
-				let new_borrow_share =
-					Percent::from_rational(amount_to_borrow, principal_after_new_borrow);
-
-				market_index
-					.safe_mul(&new_borrow_share.into())?
-					.safe_add(&account_interest_index.safe_mul(&existing_borrow_share.into())?)?
-			};
-
-			// mint debt token into user and lock it (it's used as a marker of how much the account
-			// has borrowed total)
-			<T as Config>::MultiCurrency::mint_into(
-				debt_asset_id,
-				borrowing_account,
-				amount_to_borrow,
-			)?;
-			<T as Config>::MultiCurrency::hold(debt_asset_id, borrowing_account, amount_to_borrow)?;
-
-			// transfer borrow asset from market to the borrower
-			<T as Config>::MultiCurrency::transfer(
-				borrow_asset,
-				&market_account,
-				borrowing_account,
-				amount_to_borrow,
-				false,
-			)?;
-			DebtIndex::<T>::insert(market_id, borrowing_account, new_account_interest_index);
-			BorrowTimestamp::<T>::insert(
-				market_id,
-				borrowing_account,
-				LastBlockTimestamp::<T>::get(),
-			);
-
-			if !BorrowRent::<T>::contains_key(market_id, borrowing_account) {
-				let deposit = T::WeightToFee::calc(&T::WeightInfo::liquidate(2));
-				<T as Config>::NativeCurrency::transfer(
-					borrowing_account,
-					&market_account,
-					deposit,
-					true,
-				)?;
-				BorrowRent::<T>::insert(market_id, borrowing_account, deposit);
-			} else {
-				// REVIEW
-			}
-
-			Ok(())
-		}
-
-		/// NOTE: Must be called in transaction!
-		fn repay_borrow(
-			market_id: &Self::MarketId,
-			from: &Self::AccountId,
-			beneficiary: &Self::AccountId,
-			total_repay_amount: RepayStrategy<BorrowAmountOf<Self>>,
-			keep_alive: bool,
-		) -> Result<BorrowAmountOf<Self>, DispatchError> {
-			use crate::repay_borrow::{pay_interest, repay_principal};
-
-			// cannot repay in the same block as the borrow
-			let timestamp = BorrowTimestamp::<T>::get(market_id, beneficiary)
-				.ok_or(Error::<T>::BorrowDoesNotExist)?;
-			ensure!(
-				timestamp != LastBlockTimestamp::<T>::get(),
-				Error::<T>::BorrowAndRepayInSameBlockIsNotSupported
-			);
-
-			// principal + interest
-			let beneficiary_total_debt_with_interest =
-				match Self::total_debt_with_interest(market_id, beneficiary)? {
-					TotalDebtWithInterest::Amount(amount) => amount,
-					TotalDebtWithInterest::NoDebt =>
-						return Err(Error::<T>::CannotRepayZeroBalance.into()),
-				};
-
-			let market_account = Self::account_id(market_id);
-
-			let MarketAssets { borrow_asset, debt_asset } = Self::get_assets_for_market(market_id)?;
-
-			// initial borrow amount
-			let beneficiary_borrow_asset_principal =
-				<T as Config>::MultiCurrency::balance(debt_asset, beneficiary);
-			// interest accrued
-			let beneficiary_interest_on_market = beneficiary_total_debt_with_interest
-				.safe_sub(&beneficiary_borrow_asset_principal)?;
-
-			ensure!(
-				!beneficiary_total_debt_with_interest.is_zero(),
-				Error::<T>::CannotRepayZeroBalance
-			);
-
-			let repaid_amount = match total_repay_amount {
-				RepayStrategy::TotalDebt => {
-					// pay interest, from -> market
-					// burn debt token interest from market
-					pay_interest::<T>(
-						borrow_asset,
-						debt_asset,
-						from,
-						&market_account,
-						beneficiary_interest_on_market,
-						keep_alive,
-					)?;
-
-					// release and burn debt token from beneficiary and transfer borrow asset to
-					// market, paid by `from`
-					repay_principal::<T>(
-						borrow_asset,
-						debt_asset,
-						from,
-						&market_account,
-						beneficiary,
-						beneficiary_borrow_asset_principal,
-						keep_alive,
-					)?;
-
-					beneficiary_total_debt_with_interest
-				},
-
-				// attempt to repay a partial amount of the debt, paying off interest and principal
-				// proportional to how much of each there is.
-				RepayStrategy::PartialAmount(partial_repay_amount) => {
-					ensure!(
-						partial_repay_amount <= beneficiary_total_debt_with_interest,
-						Error::<T>::CannotRepayMoreThanTotalDebt
-					);
-
-					// INVARIANT: ArithmeticError::Overflow is used as the error here as
-					// beneficiary_total_debt_with_interest is known to be non-zero at this point
-					// due to the check above (CannotRepayZeroBalance)
-
-					let interest_percentage = FixedU128::checked_from_rational(
-						beneficiary_interest_on_market,
-						beneficiary_total_debt_with_interest,
-					)
-					.ok_or(ArithmeticError::Overflow)?;
-
-					let principal_percentage = FixedU128::checked_from_rational(
-						beneficiary_borrow_asset_principal,
-						beneficiary_total_debt_with_interest,
-					)
-					.ok_or(ArithmeticError::Overflow)?;
-
-					// pay interest, from -> market
-					// burn interest (debt token) from market
-					pay_interest::<T>(
-						borrow_asset,
-						debt_asset,
-						from,
-						&market_account,
-						interest_percentage
-							.checked_mul_int::<u128>(partial_repay_amount.into())
-							.ok_or(ArithmeticError::Overflow)?
-							.into(),
-						keep_alive,
-					)?;
-
-					// release and burn debt token from beneficiary and transfer borrow asset to
-					// market, paid by `from`
-					repay_principal::<T>(
-						borrow_asset,
-						debt_asset,
-						from,
-						&market_account,
-						beneficiary,
-						principal_percentage
-							.checked_mul_int::<u128>(partial_repay_amount.into())
-							.ok_or(ArithmeticError::Overflow)?
-							.into(),
-						keep_alive,
-					)?;
-
-					// the above will short circuit if amount cannot be paid, so if this is reached
-					// then we know `partial_repay_amount` has been repaid
-					partial_repay_amount
-				},
-			};
-
-			// if the borrow is completely repaid, remove the borrow information
-			if repaid_amount == beneficiary_total_debt_with_interest {
-				// borrow no longer exists as it has been repaid in entirety, remove the
-				// timestamp & index
-				BorrowTimestamp::<T>::remove(market_id, beneficiary);
-				DebtIndex::<T>::remove(market_id, beneficiary);
-
-				// give back rent (rent = deposit)
-				let rent = BorrowRent::<T>::get(market_id, beneficiary)
-					.ok_or(Error::<T>::BorrowRentDoesNotExist)?;
-
-				<T as Config>::NativeCurrency::transfer(
-					&market_account,
-					beneficiary,
-					rent,
-					false, // we do not need to keep the market account alive
-				)?;
-			}
-
-			Ok(repaid_amount)
-		}
-
-		fn total_borrowed_from_market_excluding_interest(
-			market_id: &Self::MarketId,
-		) -> Result<Self::Balance, DispatchError> {
-			let debt_token =
-				DebtTokenForMarket::<T>::get(market_id).ok_or(Error::<T>::MarketDoesNotExist)?;
-
-			// total amount of debt *interest* owned by the market
-			let total_debt_interest =
-				<T as Config>::MultiCurrency::balance(debt_token, &Self::account_id(market_id));
-
-			let total_issued = <T as Config>::MultiCurrency::total_issuance(debt_token);
-			let total_amount_borrowed_from_market = total_issued.safe_sub(&total_debt_interest)?;
-
-			Ok(total_amount_borrowed_from_market)
-		}
-
-		fn total_interest(market_id: &Self::MarketId) -> Result<Self::Balance, DispatchError> {
-			let debt_token =
-				DebtTokenForMarket::<T>::get(market_id).ok_or(Error::<T>::MarketDoesNotExist)?;
-
-			// total amount of debt *interest* owned by the market
-			let total_debt_interest =
-				<T as Config>::MultiCurrency::balance(debt_token, &Self::account_id(market_id));
-
-			Ok(total_debt_interest)
-		}
-
-		fn accrue_interest(
-			market_id: &Self::MarketId,
-			now: Timestamp,
-		) -> Result<(), DispatchError> {
-			// we maintain original borrow principals intact on hold,
-			// but accrue total borrow balance by adding to market debt balance
-			// when user pays loan back, we reduce marked accrued debt
-			// so no need to loop over each account -> scales to millions of users
-
-			let total_borrowed_from_market_excluding_interest =
-				Self::total_borrowed_from_market_excluding_interest(market_id)?;
-			let total_available_to_be_borrowed = Self::total_available_to_be_borrowed(market_id)?;
-
-			let utilization_ratio = Self::calculate_utilization_ratio(
-				total_available_to_be_borrowed,
-				total_borrowed_from_market_excluding_interest,
-			)?;
-
-			let delta_time = now.checked_sub(LastBlockTimestamp::<T>::get()).ok_or(
-				// REVIEW: INVARIANT: this error should never happen, `now` should always
-				// be `> LastBlockTimestamp`
-				Error::<T>::Underflow,
-			)?;
-
-			let borrow_index =
-				BorrowIndex::<T>::get(market_id).ok_or(Error::<T>::MarketDoesNotExist)?;
-			let debt_asset_id =
-				DebtTokenForMarket::<T>::get(market_id).ok_or(Error::<T>::MarketDoesNotExist)?;
-
-			let accrued_interest = Markets::<T>::try_mutate(market_id, |market_config| {
-				let market_config = market_config.as_mut().ok_or(Error::<T>::MarketDoesNotExist)?;
-
-				accrue_interest_internal::<T, InterestRateModel>(
-					utilization_ratio,
-					&mut market_config.interest_rate_model,
-					borrow_index,
-					delta_time,
-					total_borrowed_from_market_excluding_interest,
-				)
-			})?;
-
-			// overwrites
-			BorrowIndex::<T>::insert(market_id, accrued_interest.new_borrow_index);
-			<T as Config>::MultiCurrency::mint_into(
-				debt_asset_id,
-				&Self::account_id(market_id),
-				accrued_interest.accrued_increment,
-			)?;
-
-			Ok(())
-		}
-
-		fn total_available_to_be_borrowed(
-			market_id: &Self::MarketId,
-		) -> Result<Self::Balance, DispatchError> {
-			let (_, market) = Self::get_market(market_id)?;
-			let borrow_asset_id = T::Vault::asset_id(&market.borrow_asset_vault)?;
-			Ok(<T as Config>::MultiCurrency::balance(borrow_asset_id, &Self::account_id(market_id)))
-		}
-
-		fn calculate_utilization_ratio(
-			cash: Self::Balance,
-			borrows: Self::Balance,
-		) -> Result<Percent, DispatchError> {
-			Ok(math::calculate_utilization_ratio(
-				LiftedFixedBalance::saturating_from_integer(cash.into()),
-				LiftedFixedBalance::saturating_from_integer(borrows.into()),
-			)?)
-		}
-
-		// previously 'borrow_balance_current'
-		fn total_debt_with_interest(
-			market_id: &Self::MarketId,
-			account: &Self::AccountId,
-		) -> Result<TotalDebtWithInterest<BorrowAmountOf<Self>>, DispatchError> {
-			let debt_token =
-				DebtTokenForMarket::<T>::get(market_id).ok_or(Error::<T>::MarketDoesNotExist)?;
-
-			// Self::get_assets_for_market()?;
-			match DebtIndex::<T>::get(market_id, account) {
-				Some(account_interest_index) => {
-					let market_interest_index =
-						BorrowIndex::<T>::get(market_id).ok_or(Error::<T>::MarketDoesNotExist)?;
-
-					let account_principal =
-						<T as Config>::MultiCurrency::balance_on_hold(debt_token, account);
-
-					if account_principal.is_zero() {
-						Ok(TotalDebtWithInterest::NoDebt)
-					} else {
-						// REVIEW
-						let account_principal =
-							LiftedFixedBalance::saturating_from_integer(account_principal.into());
-						// principal * (market index / debt index)
-						let index_ratio =
-							market_interest_index.safe_div(&account_interest_index)?;
-
-						let balance = account_principal
-							.safe_mul(&index_ratio)?
-							// TODO: Balance should be u128 eventually
-							.checked_mul_int(1_u64)
-							.ok_or(ArithmeticError::Overflow)?;
-						Ok(TotalDebtWithInterest::Amount(balance.into()))
-					}
-				},
-				None => Ok(TotalDebtWithInterest::NoDebt),
-			}
-		}
-
-		fn collateral_of_account(
-			market_id: &Self::MarketId,
-			account: &Self::AccountId,
-		) -> Result<CollateralLpAmountOf<Self>, DispatchError> {
-			AccountCollateral::<T>::get(market_id, account)
-				.ok_or_else(|| Error::<T>::MarketCollateralWasNotDepositedByAccount.into())
-		}
-
-		fn collateral_required(
-			market_id: &Self::MarketId,
-			borrow_amount: Self::Balance,
-		) -> Result<Self::Balance, DispatchError> {
-			let (_, market) = Self::get_market(market_id)?;
-			let borrow_asset = T::Vault::asset_id(&market.borrow_asset_vault)?;
-			let borrow_amount_value = Self::get_price(borrow_asset, borrow_amount)?;
-
-			Ok(LiftedFixedBalance::saturating_from_integer(borrow_amount_value.into())
-				.safe_mul(&market.collateral_factor)?
-				.checked_mul_int(1_u64)
-				.ok_or(ArithmeticError::Overflow)?
-				.into())
-		}
-
-		fn get_borrow_limit(
-			market_id: &Self::MarketId,
-			account: &Self::AccountId,
-		) -> Result<Self::Balance, DispatchError> {
-			let collateral_balance = AccountCollateral::<T>::get(market_id, account)
-				// REVIEW: I don't think this should default to zero, only to check against zero
-				// afterwards.
-				.unwrap_or_else(CollateralLpAmountOf::<Self>::zero);
-
-			if collateral_balance > T::Balance::zero() {
-				let borrower = Self::create_borrower_data(market_id, account)?;
-				let balance = borrower
-					.get_borrow_limit()
-					.map_err(|_| Error::<T>::BorrowerDataCalculationFailed)?
-					.checked_mul_int(1_u64)
-					.ok_or(ArithmeticError::Overflow)?;
-				Ok(balance.into())
-			} else {
-				Ok(Self::Balance::zero())
-			}
-		}
-	}
-
-	// various helper functions
-
-	/// given collateral information, how much of borrow asset can get?
-	pub fn swap(
-		collateral_balance: &LiftedFixedBalance,
-		collateral_price: &LiftedFixedBalance,
-		collateral_factor: &MoreThanOneFixedU128,
-	) -> Result<LiftedFixedBalance, ArithmeticError> {
-		collateral_balance.safe_mul(collateral_price)?.safe_div(collateral_factor)
-	}
-
-	/// ```python
-	/// delta_interest_rate = delta_time / period_interest_rate
-	/// debt_delta = debt_principal * delta_interest_rate
-	/// new_accrued_debt = accrued_debt + debt_delta
-	/// total_debt = debt_principal + new_accrued_debt
-	/// ```
-	pub(crate) fn accrue_interest_internal<T: Config, I: InterestRate>(
-		utilization_ratio: Percent,
-		interest_rate_model: &mut I,
-		borrow_index: OneOrMoreFixedU128,
-		delta_time: DurationSeconds,
-		total_borrows: T::Balance,
-	) -> Result<AccruedInterest<T>, DispatchError> {
-		let total_borrows: FixedU128 =
-			FixedU128::checked_from_integer(Into::<u128>::into(total_borrows))
-				.ok_or(ArithmeticError::Overflow)?;
-
-		let borrow_rate = interest_rate_model
-			.get_borrow_rate(utilization_ratio)
-			.ok_or(Error::<T>::BorrowRateDoesNotExist)?;
-
-		// borrow_rate * index * delta_time / SECONDS_PER_YEAR_NAIVE + index
-		let borrow_rate_delta = borrow_rate
-			.safe_mul(&FixedU128::saturating_from_integer(delta_time))?
-			.safe_div(&FixedU128::saturating_from_integer(SECONDS_PER_YEAR_NAIVE))?;
-
-		let new_borrow_index =
-			borrow_rate_delta.safe_mul(&borrow_index)?.safe_add(&borrow_index)?;
-
-		let accrued_increment = total_borrows
-			.safe_mul(&borrow_rate_delta)?
-			.checked_mul_int(1_u64)
-			.ok_or(ArithmeticError::Overflow)?
-			.into();
-
-		Ok(AccruedInterest { accrued_increment, new_borrow_index })
-	}
-
-	#[derive(Debug, PartialEqNoBound)]
-	pub(crate) struct AccruedInterest<T: Config> {
-		pub(crate) accrued_increment: T::Balance,
-		pub(crate) new_borrow_index: FixedU128,
-	}
-
-	/// Retrieve the current interest rate for the given `market_id`.
-	pub fn current_interest_rate<T: Config>(market_id: MarketId) -> Result<Rate, DispatchError> {
-		let market_id = MarketIndex::new(market_id);
-		let total_borrowed_from_market_excluding_interest =
-			Pallet::<T>::total_borrowed_from_market_excluding_interest(&market_id)?;
-		let total_available_to_be_borrowed =
-			Pallet::<T>::total_available_to_be_borrowed(&market_id)?;
-		let utilization_ratio = Pallet::<T>::calculate_utilization_ratio(
-			total_available_to_be_borrowed,
-			total_borrowed_from_market_excluding_interest,
-		)?;
-
-		Markets::<T>::try_get(market_id)
-			.map_err(|_| Error::<T>::MarketDoesNotExist)?
-			.interest_rate_model
-			.get_borrow_rate(utilization_ratio)
-			.ok_or(Error::<T>::BorrowRateDoesNotExist)
-			.map_err(Into::into)
 	}
 }
